add roi layers; change default config
