+++ dataset/models/tf/layers/__init__.py
@@ -6,3 +6,4 @@ from .core import flatten, flatten2d, maxout, mip
 from .conv_block import conv_block
 from .conv import conv1d_transpose, conv_transpose, separable_conv, separable_conv1d, separable_conv2d, separable_conv3d
 from .pooling import max_pooling, average_pooling, global_average_pooling, global_max_pooling, fractional_max_pooling
+from .roi import non_max_suppression, roi_pooling_layer
